update: inline announce processing job into announce
